refactor: 루트 패키지명 javachess로 변경
