--- src/main/java/chess/JavaChessApplication.java
@@ -1,13 +1,0 @@
-package chess;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class JavaChessApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(JavaChessApplication.class, args);
-    }
-
-}
