--- src/main/java/chess/chess/ChessStrings.java
@@ -1,21 +1,0 @@
-package chess.chess;
-
-public class ChessStrings {
-
-    public static final String WHITE_KING = "♔";
-    public static final String WHITE_QUEEN = "♕";
-    public static final String WHITE_ROOK = "♖";
-    public static final String WHITE_BISHOP = "♗";
-    public static final String WHITE_KNIGHT = "♘";
-    public static final String WHITE_PAWN = "♙";
-    public static final String BLACK_KING = "♚";
-    public static final String BLACK_QUEEN = "♛";
-    public static final String BLACK_ROOK = "♜";
-    public static final String BLACK_BISHOP = "♝";
-    public static final String BLACK_KNIGHT = "♞";
-    public static final String BLACK_PAWN = "♟";
-
-    private ChessStrings() {
-    }
-
-}
