--- src/main/java/chess/chess/StringUtils.java
@@ -1,10 +1,0 @@
-package chess.chess;
-
-public class StringUtils {
-
-    public static final String NEWLINE = System.lineSeparator();
-
-    private StringUtils() {
-    }
-
-}
