+++ src/main/java/javachess/JavaChessApplication.java
@@ -1,0 +1,13 @@
+package javachess;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class JavaChessApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(JavaChessApplication.class, args);
+    }
+
+}
