+++ src/main/java/javachess/chess/BoardContext.java
@@ -1,0 +1,10 @@
+package javachess.chess;
+
+import javachess.chess.piece.Piece;
+
+public interface BoardContext {
+
+    boolean isEmptyAt(Offset offset);
+
+    boolean isColorAt(Offset offset, Piece.Color color);
+}
