+++ src/main/java/javachess/chess/StringUtils.java
@@ -1,0 +1,10 @@
+package javachess.chess;
+
+public class StringUtils {
+
+    public static final String NEWLINE = System.lineSeparator();
+
+    private StringUtils() {
+    }
+
+}
