--- src/test/java/chess/JavaChessApplicationTests.java
@@ -1,13 +1,0 @@
-package chess;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class JavaChessApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
