+++ src/test/java/javachess/JavaChessApplicationTests.java
@@ -1,0 +1,13 @@
+package javachess;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class JavaChessApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
